fix: Add steam API route
