+++ packages/steam/SteamAuthComponent.cs
@@ -1,6 +1,5 @@
 using SteamKit2;
 using SteamKit2.Authentication;
-using System.Collections.Concurrent;
 
 namespace Steam
 {
@@ -12,35 +11,33 @@ namespace Steam
         private readonly CallbackManager _manager;
         private readonly string _userId;
         private readonly Action<string, string>? _onCredentialsObtained;
-        private readonly IServiceProvider _serviceProvider;
         private CancellationTokenSource? _cts;
         private Task? _callbackTask;
         private TaskCompletionSource<bool>? _authCompletionSource;
         private QrAuthSession? _authSession;
-        
+
         private string? _username;
         private string? _refreshToken;
         private bool _isLoggedIn = false;
-        
+
         // Event that clients can subscribe to
         public event Action<ServerSentEvent>? OnEvent;
-        
+
         // Store account information
         private readonly Dictionary<string, object> _accountInfo = new();
 
-        public SteamAuthComponent(IServiceProvider serviceProvider, string userId, Action<string, string>? onCredentialsObtained = null)
+        public SteamAuthComponent(string userId, Action<string, string>? onCredentialsObtained = null)
         {
             _userId = userId;
             _onCredentialsObtained = onCredentialsObtained;
-            _serviceProvider = serviceProvider;
-            
+
             var configuration = SteamConfig.GetDefaultSteamClientConfig();
             // Create SteamKit2 client
             _steamClient = new SteamClient(configuration);
             _manager = new CallbackManager(_steamClient);
             _steamUser = _steamClient.GetHandler<SteamUser>() ?? throw new InvalidOperationException("SteamUser handler not available");
             _steamFriends = _steamClient.GetHandler<SteamFriends>() ?? throw new InvalidOperationException("SteamFriends handler not available");
-            
+
             // Register callbacks
             _manager.Subscribe<SteamClient.ConnectedCallback>(OnConnected);
             _manager.Subscribe<SteamClient.DisconnectedCallback>(OnDisconnected);
@@ -54,16 +51,16 @@ namespace Steam
         public Action Subscribe(Action<ServerSentEvent> callback)
         {
             OnEvent += callback;
-            
+
             // If we already have a QR code URL, send it immediately
             if (_authSession != null)
             {
                 callback(new ServerSentEvent("challenge_url", new { url = _authSession.ChallengeURL }));
             }
-            
+
             return () => OnEvent -= callback;
         }
-        
+
         public void SetCredentials(string username, string refreshToken)
         {
             _username = username;
@@ -76,7 +73,7 @@ namespace Steam
             {
                 return _isLoggedIn; // Already connected
             }
-            
+
             _authCompletionSource = new TaskCompletionSource<bool>();
             _cts = new CancellationTokenSource();
 
@@ -102,14 +99,14 @@ namespace Steam
         public void Disconnect()
         {
             _cts?.Cancel();
-            
+
             if (_isLoggedIn)
             {
                 _steamUser.LogOff();
             }
-            
+
             _steamClient.Disconnect();
-            
+
             NotifyEvent(new ServerSentEvent("disconnected", new { success = true }));
         }
 
@@ -119,7 +116,7 @@ namespace Steam
         }
 
         #region Steam Callbacks
-        
+
         private async void OnConnected(SteamClient.ConnectedCallback callback)
         {
             try
@@ -130,7 +127,7 @@ namespace Steam
                 if (!string.IsNullOrEmpty(_refreshToken) && !string.IsNullOrEmpty(_username))
                 {
                     NotifyEvent(new ServerSentEvent("status", new { message = $"Logging in as '{_username}'..." }));
-                    
+
                     _steamUser.LogOn(new SteamUser.LogOnDetails
                     {
                         Username = _username,
@@ -141,34 +138,34 @@ namespace Steam
                 {
                     // Start QR authentication flow
                     NotifyEvent(new ServerSentEvent("status", new { message = "Starting QR authentication..." }));
-                    
+
                     _authSession = await _steamClient.Authentication.BeginAuthSessionViaQRAsync(
                         new AuthSessionDetails { DeviceFriendlyName = "Steam Auth Client" });
-                    
+
                     // Handle URL changes
                     _authSession.ChallengeURLChanged = () =>
                     {
                         NotifyEvent(new ServerSentEvent("status", new { message = "QR Code link changed" }));
                         NotifyEvent(new ServerSentEvent("challenge_url", new { url = _authSession.ChallengeURL }));
                     };
-                    
+
                     // Send initial QR code URL
                     NotifyEvent(new ServerSentEvent("challenge_url", new { url = _authSession.ChallengeURL }));
-                    
+
                     // Poll for authentication result
                     await Task.Run(async () =>
                     {
                         try
                         {
                             var pollResponse = await _authSession.PollingWaitForResultAsync();
-                            
+
                             // Store credentials
                             _username = pollResponse.AccountName;
                             _refreshToken = pollResponse.RefreshToken;
-                            
+
                             // Log in with obtained credentials
                             NotifyEvent(new ServerSentEvent("status", new { message = $"Logging in as '{_username}'..." }));
-                            
+
                             _steamUser.LogOn(new SteamUser.LogOnDetails
                             {
                                 Username = pollResponse.AccountName,
@@ -196,7 +193,7 @@ namespace Steam
         {
             NotifyEvent(new ServerSentEvent("status", new { message = "Disconnected from Steam" }));
             _isLoggedIn = false;
-            
+
             // Only try to reconnect if not deliberately disconnected
             if (_callbackTask != null && !_cts!.IsCancellationRequested)
             {
@@ -221,25 +218,25 @@ namespace Steam
 
             _isLoggedIn = true;
             NotifyEvent(new ServerSentEvent("status", new { message = "Successfully logged in to Steam" }));
-            
+
             // Store basic info
             if (callback.ClientSteamID != null)
             {
                 _accountInfo["steamId"] = callback.ClientSteamID.ConvertToUInt64();
             }
-            
+
             NotifyEvent(new ServerSentEvent("login-success", new
             {
                 steamId = _steamUser.SteamID?.ConvertToUInt64(),
                 username = _username
             }));
-            
+
             // Save credentials if callback provided
             if (_onCredentialsObtained != null && !string.IsNullOrEmpty(_refreshToken) && !string.IsNullOrEmpty(_username))
             {
                 _onCredentialsObtained(_username, _refreshToken);
             }
-            
+
             // Request persona state
             if (_steamUser.SteamID != null)
             {
@@ -252,24 +249,24 @@ namespace Steam
             _isLoggedIn = false;
             NotifyEvent(new ServerSentEvent("logged-off", new { reason = callback.Result.ToString() }));
         }
-        
+
         private void OnAccountInfo(SteamUser.AccountInfoCallback callback)
         {
             _accountInfo["country"] = callback.Country;
             _accountInfo["personaName"] = callback.PersonaName;
-            
+
             // Request additional user info
             if (_steamFriends != null && _steamUser.SteamID != null)
             {
                 _steamFriends.RequestFriendInfo(_steamUser.SteamID);
             }
         }
-        
+
         private void OnEmailInfo(SteamUser.EmailAddrInfoCallback callback)
         {
             _accountInfo["email"] = callback.EmailAddress;
         }
-        
+
         private void OnPersonaState(SteamFriends.PersonaStateCallback callback)
         {
             // Only care about our own persona state
@@ -282,7 +279,7 @@ namespace Steam
                     var avatarUrl = $"https://avatars.akamai.steamstatic.com/{avatarStr}_full.jpg";
                     _accountInfo["avatarUrl"] = avatarUrl;
                 }
-                
+
                 // Update account info
                 _accountInfo["username"] = _username ?? "Unknown";
                 _accountInfo["personaName"] = callback.Name;
@@ -291,15 +288,15 @@ namespace Steam
                 _accountInfo["lastLogOn"] = callback.LastLogOn;
                 _accountInfo["lastLogOff"] = callback.LastLogOff;
                 _accountInfo["steamId"] = _steamUser.SteamID.ConvertToUInt64();
-                
+
                 // Notify client of the updated account info
                 NotifyEvent(new ServerSentEvent("account_info", new { info = _accountInfo }));
-                
+
                 // Complete authentication process
                 _authCompletionSource?.TrySetResult(true);
             }
         }
-        
+
         #endregion
     }
 }
