+++ packages/steam/SteamConfiguration.cs
@@ -1,0 +1,17 @@
+using SteamKit2;
+
+namespace Steam
+{
+    public class SteamConfig
+    {
+        public static SteamConfiguration GetDefaultSteamClientConfig()
+        {
+            return SteamConfiguration.Create(config =>
+               {
+                   config.WithConnectionTimeout(TimeSpan.FromSeconds(30));
+                   config.WithProtocolTypes(ProtocolTypes.Tcp | ProtocolTypes.WebSocket);
+               }
+            );
+        }
+    }
+}
